32 remove registration link remove language from main page add language to login page
