+++ src/pages/home/home.ts
@@ -5,9 +5,6 @@ import { RestProvider } from '../../providers/rest/rest';
 import { LoadingController } from 'ionic-angular';
 import { Geolocation, GeolocationOptions, Geoposition, PositionError } from '@ionic-native/geolocation';
 
-import { TranslateService } from '@ngx-translate/core';
-import { LanguageServiceProvider } from "../../providers/language-service/language-service";
-import { LanguageModel } from "../../models/language.model";
 
 @IonicPage({
   name: 'home'
@@ -19,8 +16,6 @@ import { LanguageModel } from "../../models/language.model";
 })
 export class HomePage {
 
-  languageSelected: any = 'en';
-  languages: Array<LanguageModel>;
 
   public stories = new Array();
   public posts = new Array();
@@ -49,9 +44,7 @@ export class HomePage {
     public navCtrl: NavController,
     public restProvider: RestProvider,
     public loadingCtrl: LoadingController,
-    private geolocation: Geolocation,
-    public translate: TranslateService,
-    public languageService: LanguageServiceProvider
+    private geolocation: Geolocation
   ) {
 
     this.detailPage = PlaygroundDetailPage;
@@ -59,8 +52,6 @@ export class HomePage {
     let loader = loadingCtrl.create({ content: "در حال بارگذاری ..." });
     loader.present();
 
-    this.languages = this.languageService.getLanguages();
-    this.setLanguage();
 
     restProvider.getStories(0).subscribe(stories => {
       console.log('stories : ', stories);
@@ -128,17 +119,6 @@ export class HomePage {
     }, 1000);
   }
 
-  setLanguage() {
-    let defaultLanguage = this.translate.getDefaultLang();
-    this.languageSelected="fa";
-    if (this.languageSelected) {
-      this.translate.setDefaultLang(this.languageSelected);
-      this.translate.use(this.languageSelected);
-    } else {
-      this.languageSelected = defaultLanguage;
-      this.translate.use(defaultLanguage);
-    }
-  }
 
   likeButton() {
     if (this.like_btn.icon_name === 'heart-outline') {
