+++ src/pages/login/login.html
@@ -1,18 +1,38 @@
-<!--
-  Generated template for the LoginPage page.
-
-  See http://ionicframework.com/docs/components/#navigation for more info on
-  Ionic pages and navigation.
--->
-<ion-header>
-
-  <ion-navbar>
-    <ion-title>login</ion-title>
-  </ion-navbar>
-
-</ion-header>
-
-
-<ion-content padding>
-
+<ion-content padding class="transparent-header">
+  <ion-header>
+    <ion-navbar>
+    </ion-navbar>
+  </ion-header>
+  <img class="logo" src="http://masjedcloob.ir/img/masjedCloob_02.png" />
+  <div padding>
+    <ion-item>
+      <ion-input type="text" placeholder="موبایل"></ion-input>
+    </ion-item>
+    <ion-item>
+      <ion-input type="password" placeholder="کد 4 رقمی"></ion-input>
+    </ion-item>
+    <button ion-button block outline (click)="login()" color="light" class="login-button">ورود</button>
+  </div>
+  <div class="strike">
+    <span>یا</span>
+  </div>
+  <button ion-button block clear (click)="login()" color="light" class="login-button">
+    <ion-icon name="logo-eeta"></ion-icon> ورود با فیس بوک
+  </button>
 </ion-content>
+<ion-footer>
+  <ion-toolbar class="footer">
+    <!-- <div (click)="goToSignup()">
+      <span>نام کاربری نساخته اید؟ <strong>از اینجا عضو شوید</strong>.</span>
+    </div> -->
+    <ion-list no-lines>
+      <ion-item>
+        <ion-label>{{ 'SELECT_LANGUAGE' | translate }}</ion-label>
+        <ion-select [(ngModel)]="languageSelected" (ionChange)='setLanguage()' [cancelText]="'CANCEL' | translate"
+          [okText]="'OK' | translate">
+          <ion-option *ngFor="let item of languages" [value]="item.code">{{item.name}}</ion-option>
+        </ion-select>
+      </ion-item>
+    </ion-list>
+  </ion-toolbar>
+</ion-footer>
