+++ src/pages/login/login.scss
@@ -1,3 +1,117 @@
 page-login {
-
-}
+    $transparent-white: rgba(255, 255, 255, .6);
+    // This is just to make the header transparent and button/icon with white color
+    .toolbar-content,
+    span {
+      color: white;
+    }
+    .toolbar-background,
+    ion-navbar,
+    ion-header {
+      background: transparent !important;
+      background-color: transparent !important;
+    }
+    .header-md::after,
+    .toolbar-background-ios {
+      background-image: none;
+      border-bottom: 0;
+    }
+    // Footer
+    ion-footer {
+      height: 47px;
+    }
+    .footer {
+      text-align: center;
+      background: rgba(255, 255, 255, .06);
+      min-height: 47px;
+    }
+    .footer-md::before {
+      border-top: 1px solid $transparent-white;
+      background-image: none;
+    }
+    // Fullscreen animated rainbow gradient
+    ion-content {
+      background: linear-gradient(316deg, #ff2400, #e81d1d, #e8b71d, #e3e81d, #1de840, #1ddde8, #2b1de8, #dd00f3, #dd00f3);
+      background-size: 1800% 1800%;
+      animation: rainbow 30s linear infinite;
+    }
+    @keyframes rainbow {
+      0% {
+        background-position: 0% 0%
+      }
+      50% {
+        background-position: 100% 100%
+      }
+      100% {
+        background-position: 0% 0%
+      }
+    }
+    // Content
+    .scroll-content {
+      display: flex;
+      flex-direction: column;
+      overflow: hidden;
+    }
+    .logo {
+      width: 50%;
+      height: auto;
+      margin: 0 auto;
+      padding-top: 50px;
+    }
+    // Input items
+    ion-item {
+      border-radius: 3px !important;
+      padding-left: 10px !important;
+      margin-bottom: 15px !important;
+      margin-bottom: 10px;
+      color: #fff !important;
+      background-color: rgba(255, 255, 255, .2) !important;
+      font-size: 0.9em;
+    }
+     ::-webkit-input-placeholder {
+      color: white !important;
+    }
+    .item-inner {
+      border-bottom: 0 !important;
+      box-shadow: none !important;
+    }
+    // --
+    .login-button {
+      text-transform: none;
+      border-radius: 3px !important;
+    }
+    // eeta icon
+    ion-icon {
+      padding-right: 10px;
+      font-size: 2.2rem;
+    }
+    // HR with text in the middle effect
+    .strike {
+      display: block;
+      text-align: center;
+      overflow: hidden;
+      white-space: nowrap;
+    }
+    .strike>span {
+      position: relative;
+      display: inline-block;
+      font-weight: bold;
+    }
+    .strike>span:before,
+    .strike>span:after {
+      content: "";
+      position: absolute;
+      top: 50%;
+      width: 9999px;
+      height: 1px;
+      background: $transparent-white;
+    }
+    .strike>span:before {
+      right: 100%;
+      margin-right: 15px;
+    }
+    .strike>span:after {
+      left: 100%;
+      margin-left: 15px;
+    }
+  }
