+++ src/pages/login/login.ts
@@ -1,12 +1,10 @@
 import { Component } from '@angular/core';
-import { IonicPage, NavController, NavParams } from 'ionic-angular';
+import { IonicPage, NavController, NavParams, LoadingController } from 'ionic-angular';
+import { TabsPage } from '../tabs/tabs';
+import { TranslateService } from '@ngx-translate/core';
+import { LanguageServiceProvider } from "../../providers/language-service/language-service";
+import { LanguageModel } from "../../models/language.model";
 
-/**
- * Generated class for the LoginPage page.
- *
- * See https://ionicframework.com/docs/components/#navigation for more info on
- * Ionic pages and navigation.
- */
 
 @IonicPage()
 @Component({
@@ -15,11 +13,47 @@ import { IonicPage, NavController, NavParams } from 'ionic-angular';
 })
 export class LoginPage {
 
-  constructor(public navCtrl: NavController, public navParams: NavParams) {
+  
+  languageSelected: any ;
+  languages: Array<LanguageModel>;
+
+  constructor(public navCtrl: NavController,
+    public loadingCtrl: LoadingController,
+    public translate: TranslateService,
+    public languageService: LanguageServiceProvider,
+    public navParams: NavParams) {
+      
+    this.languages = this.languageService.getLanguages();
+    this.setLanguage();
+
   }
 
   ionViewDidLoad() {
     console.log('ionViewDidLoad LoginPage');
   }
 
+
+  login() {
+    const loading = this.loadingCtrl.create({
+      duration: 500
+    });
+
+    loading.onDidDismiss(() => {
+      this.navCtrl.setRoot(TabsPage);
+    });
+
+    loading.present();
+
+  }
+
+  setLanguage() {
+    let defaultLanguage = this.translate.getDefaultLang();
+    if (this.languageSelected) {
+      this.translate.setDefaultLang(this.languageSelected);
+      this.translate.use(this.languageSelected);
+    } else {
+      this.languageSelected = defaultLanguage;
+      this.translate.use(defaultLanguage);
+    }
+  }
 }
